build system: size_t and icu-config version 4.4
git-svn-id: svn://tug.org/texlive/trunk/Build/source@17609 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/kpathsea/xcalloc.c
@@ -20,14 +20,14 @@
 
 
 void *
-xcalloc (unsigned nelem,  unsigned elsize)
+xcalloc (size_t nelem,  size_t elsize)
 {
     void *new_mem = (void*)calloc(nelem ? nelem : 1, elsize ? elsize : 1);
   
     if (new_mem == NULL) {
         fprintf(stderr,
-                "xcalloc: request for %u elements of size %u failed.\n",
-                nelem, elsize);
+                "xcalloc: request for %lu elements of size %lu failed.\n",
+                (unsigned long)nelem, (unsigned long)elsize);
         exit(EXIT_FAILURE);
     }
